requirements.txt
xmltodict==0.12.0
chart-studio==1.1.0
python-igraph==0.9.1